add trashable container & some style changes in group modal
