+++ app/src/pages/AppsPage/AppsPageContent.tsx
@@ -14,7 +14,7 @@ import { useOpenModalSearchParams } from '@/hooks/modal'
 import { createPortal } from 'react-dom'
 import { restrictToWindowEdges } from '@dnd-kit/modifiers'
 import { PinsGroup } from './components/PinsGroup'
-import { PinGroupModal } from './components/PinGroupModal'
+import { PinGroupModal } from './components/PinGroupModal/PinGroupModal'
 import { useSensors } from './utils/useSensors'
 import { usePinDragAndDrop } from './utils/usePinDragAndDrop'
 import { AppNostro } from '@/shared/AppNostro/AppNostro'
