+++ app/src/pages/AppsPage/components/PinGroupModal/AppNostroRemovable.tsx
